Fix warning (64-bit)
git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/turbine/trunk@5693 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/tcl/blob/blob.c
@@ -47,7 +47,7 @@ SwiftBlob_sizeof_float(void)
 void*
 SwiftBlob_cast_to_ptr(int i)
 {
-  return (void*) i;
+  return (void*) (size_t)i;
 }
 
 int
